changed 'addresses' accordion to 'contact information', moved phone number, email address, contact name there
