+++ src/components/ViewDirectoryEntry/sectionsShared/index.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as DirectoryEntryInfo } from './DirectoryEntryInfo';
-export { default as Addresses } from './Addresses';
+export { default as ContactInformation } from './ContactInformation';
 export { default as Services } from './Services';
 export { default as CustomProperties } from './CustomProperties';
