+++ translations/ui-directory/en_US.json
@@ -46,6 +46,9 @@
   "information.phoneNumber": "Phone number",
   "information.emailAddress": "Email address",
   "information.contactName": "Contact name",
+  "information.mainPhoneNumber": "Main phone number",
+  "information.mainEmailAddress": "Main email address",
+  "information.mainContactName": "Main contact name",
   "information.description": "Description",
   "information.symbols": "Symbols",
   "information.lmsLocationCode": "LMS location code",
@@ -60,6 +63,7 @@
   "information.members": "Members",
   "information.tags": "Tags",
   "information.heading.addresses": "Addresses",
+  "information.heading.contactInformation": "Contact information",
   "information.addressNofM": "Address {index} of {count}",
   "information.heading.services": "Services",
   "information.serviceNofM": "Service {index} of {count}",
